flask_db/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
python-dotenv==0.19.0
mysql-connector-python==8.0.26
Werkzeug==2.0.1
email-validator==1.1.3